requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
gunicorn==21.2.0
eventlet==0.33.3
earthengine-api==0.1.384
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.15.0
rasterio==1.3.9
geopandas==0.14.1
shapely==2.0.2
matplotlib==3.8.2
pillow==10.1.0
python-dotenv==1.0.0
geopy==2.4.1
joblib==1.3.2
urllib3==2.0.7
